% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management_month.R
\name{commision_management_monthServer}
\alias{commision_management_monthServer}
\title{处理逻辑}
\usage{
commision_management_monthServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
commision_management_monthServer()
}
