% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_managementupdateServer}
\alias{grossProfit_managementupdateServer}
\title{处理逻辑}
\usage{
grossProfit_managementupdateServer(
  input,
  output,
  session,
  dms_token,
  erp_token
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}

\item{erp_token}{}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
grossProfit_managementServer()
}
