% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_detailAmt_actual_year.R
\name{rebate_detailAmt_actual_yearServer}
\alias{rebate_detailAmt_actual_yearServer}
\title{处理逻辑}
\usage{
rebate_detailAmt_actual_yearServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
rebate_detailAmt_actual_yearServer()
}
