% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{rebate_exchangerateServer}
\alias{rebate_exchangerateServer}
\title{处理逻辑}
\usage{
rebate_exchangerateServer(input, output, session, dms_token, erp_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
rebate_exchangerateServer()
}
