% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_management_month.R
\name{rebate_management_monthServer}
\alias{rebate_management_monthServer}
\title{处理逻辑}
\usage{
rebate_management_monthServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
rebate_management_monthServer()
}
